Components - steps-form/components - preventing inmediate form submision
At this commit we added a local state called ( submit ) to change form
button type. Also, we added some delay using useEffect looking for
step id changes and makes changes to the local state. The final goal
is to prevent the inmediate form submission when the summary step is mounted.